requirements.txt
# Core dependencies
python-dotenv>=0.19.0
pytz>=2021.1
